src/com/incluctab/tddClasses/chapterThree/HealthProfile.java
package com.incluctab.tddClasses.chapterThree;

import java.time.Year;

public class HealthProfile {

    private String firstName;
    private String lastName;
    private String gender;
    private Date dateOfBirth;
    private double height;
    private double weight;

    public HealthProfile(String firstName, String lastName, String gender, Date dateOfBirth, double height, double weight) {
        this.firstName = firstName;
        this.lastName = lastName;
        this.gender = gender;
        this.dateOfBirth = dateOfBirth;
        this.height = height;
        this.weight = weight;
    }

    public  String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public String getGender() {
        return gender;
    }

    public void setGender(String gender) {
        this.gender = gender;
    }

    public Date getDateOfBirth() {
        return dateOfBirth;
    }

    public void setDateOfBirth(Date dateOfBirth) {
        this.dateOfBirth = dateOfBirth;
    }

    public Double getHeight() {
        return height;
    }

    public void setHeight(double height) {
        this.height = height;
    }

    public double getWeight() {
        return weight;
    }

    public void setWeight(double weight) {
        this.weight = weight;
    }

    public int calculateAge(int year) {
        int age = Year.now().getValue() - year;
        return age;
    }
}
